+++ mywords-go/cmd/main.go
@@ -3,10 +3,11 @@
 package main
 
 import (
+	"embed"
 	"flag"
 	"fmt"
+	"io/fs"
 	"mywords/mylog"
-	"mywords/www"
 	"net"
 	"net/http"
 	"os"
@@ -14,6 +15,9 @@ import (
 	"runtime"
 	"time"
 )
+
+//go:embed web/*
+var webEmbed embed.FS
 
 func main() {
 
@@ -37,7 +41,7 @@ func main() {
 	mux.HandleFunc("/_downloadBackUpdate", downloadBackUpdate)
 	mux.HandleFunc("/_webParseAndSaveArticleFromFile", webParseAndSaveArticleFromFile)
 	mux.HandleFunc("/_webRestoreFromBackUpData", webRestoreFromBackUpData)
-	mux.Handle("/", http.FileServer(www.FileSystem))
+	mux.Handle("/", http.FileServer(http.FS(&webEmbedHandler{webEmbed: webEmbed})))
 	mylog.Info("server start", "port", *port, "rootDir", *rootDir)
 	go func() {
 		time.Sleep(time.Second)
@@ -65,4 +69,14 @@ func getApplicationDir() (string, error) {
 	// 请注意，如果同时打开多个应用，可能会导致目录冲突，数据造成不一致或丢失
 	defaultRootDir = filepath.ToSlash(defaultRootDir)
 	return defaultRootDir, err
+}
+
+type webEmbedHandler struct {
+	webEmbed embed.FS
+}
+
+func (f webEmbedHandler) Open(name string) (fs.File, error) {
+	// 在windows系统下必须用toSlash 封装一下路径，否则，web\index.html!=web/index.html
+	name = filepath.ToSlash(filepath.Join("web", name))
+	return f.webEmbed.Open(name)
 }
